CREDITS/CONTACT: remove references to namesys.com
While Namesys drove almost all reiserfs development and was instrumental
in supporting it for many years, it has been offline for the past
6 years. All of its founders and employees have moved on to other
endeavors or are otherwise unavailable. This can be confusing to users
since the documentation points there either for background information
or to have bugs fixed.

This patch removes most of the non-historical references to
namesys.com and directs users to the reiserfs-devel list on vger.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>